refactor(lib): extract inline test module to src/tests.rs
Move the 1,974-line #[cfg(test)] mod tests { ... } block (lines 4402-6375)
from crates/aptu-coder/src/lib.rs into a new file
crates/aptu-coder/src/tests.rs. Replace the inline block in lib.rs with
#[cfg(test)] mod tests;.

No logic changes, no public API changes. All 68 inline tests continue to
pass after the move.

Closes #1193
